+++ src/component/index.js
@@ -1,4 +1,4 @@
-import tree from './element/packages/tree/tree.vue';
+import tree from './element/packages/tree/src/tree.vue';
 import areaTree from './area-tree/area-tree.vue';
 
 export {tree, areaTree};